feat: add singUp function on authContext
To use it when an user is creating an account but needs to choose upload a profile picture or not. with the singup function we're setting the user on the context but not saving it on the local storage, so we can call the login functon on the buttons in the view of upload a profile picture right after create an account.